Implement cache server
Cache server allows balancer instances to fetch value for a
key from others. It means that if one instance does not have
a value for a key it might ask others to find out if anyone has
one.

Interesting usage of Sync + Send traits which mean
that both `InstanceHost` and `CacheServer` implementations
must allow moving the data from one thread to another.

https://doc.rust-lang.org/nomicon/send-and-sync.html